requirements.txt
Flask
gunicorn
google-genai
google-cloud-secret-manager
firabase-admin
